server/requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.4
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
pytest==8.1.1
djangorestframework-simplejwt
